cls/rgw: fixup range for encrypted object in prefetch_head

+++ src/rgw/rgw_basic_types.h
@@ -627,7 +627,7 @@ inline std::ostream& operator<<(std::ostream& out, const rgw_user &u) {
   return out << s;
 }
 struct prefetch_range {
-  size_t off = 0;
+  off_t off = 0;
   uint64_t len = std::numeric_limits<uint64_t>::max();
 };
 #endif
